+++ CHANGELOG.rst
@@ -12,6 +12,10 @@ Fixed
 ^^^^^
 - ``data``: Removed pylandsat dependency and added methods for downloading landsat products from GCS in data module #106
 
+Added
+^^^^^
+- ``data``: Streaming for Scihub #138
+
 [1.2.1] (2021-05-05)
 ---------------------